Major reworks to Gameplay
Added Atmos Massive Clouds System (external tool).
Created a class that handles the Player moves' feedback during the battle.
Created a class that handles the Enemy moves' feedback during the battle.
Created a class that handles the different VFX on the Battlefield.
Created a class that manages the Rewards at the end of a Battle.
Created a class to save the Player Stats.
Created a class to handle the Player stats.
Created a class to handle the Summoning System during a battle.
Test - Created a Spell Scriptable Object
Iterated on BattleManager class to accomodate the new systems.
Streamlined existing logic in the BattleManager relying on event-driven architecture.
Miscellaneous.