Summon Mechanic - Feature Development
Added methods and did groundwork for Summon Mechanic.
Now a button spawns in the Active Player Menu allowing to Summon the Deity. The Deity actually appears on the battlefield.
Laid groundwork for a system that checks a parameter (i.e.: killed enemies) and unlocks the possibility of fighting against a Deity.