Move and Melee Player Action Development
Corrected a bug that prevented the Player Unit to save the new position correctly after moving resulting in an distance calculation error in the Melee Attack that ultimately led to broke the mechanic.
Now the Player can move, attack the Enemy Unit and apply the Knockback as intended (closer to Enemy -> stronger knockback).