Summon Deity - Feature Development
Now the Player can successfully Summon a Deity on the Battlefield using the Summon Button. Then, the Summon Button switches to Pray and the Player can use it to increase the Deity Power (charging their attack).
Note: Encountered a bug that allows the Summon to Spawn multiple times. Will need to fix later.