Spell VFX Implementation + Move Name announcement
Implemented VFX (from a package) into the project. Now the Player can use a Single Target or AOE Spell and the corresponding feedback plays accordingly. Player Unit Template Scriptable Objects hold a reference to the VFX.
When the Player Unit uses a move, now the move of the name and the user appear on the central button Move Name panel.
Note: I will need later to add a system that clears the space or adds a generic sentence (Pokémon Style).