Feature Development - Deity Link
Now the Player can capture a Deity and the link is persistent in the memory. I will later need to further flesh out the logic and the UI in a way that only available Deities appear on the Captured Deities menu. Also, I need to rename the Game Manager Captured Deity List to something more fitting with the nature of the system (e.g.: Deity Pool).