Clear Missing Tiles in Deity Battle
Hotfix: I'm using a method to clear Tiles with Missing Units at Update.
In the next updates, I need to find the exact point in the execution stack where the Grid receives the Enemy coords and sets their tiles as Occupied (even with a Missing Enemy GameObject) and add a check to prevent this behaviour.