Handle cases where the Player clicks on Empty Tiles during Melee
Adjusted the Select Logic in Melee Player Action. Now the Player receives a negative feedback (glowing red tile) when selecting a Tile with no Enemy Unit on it. Also fixed the previous Null Ref error (since the method now checks for the existence of a currentTarget before executing the selection logic).