Update: Player Units profiles
Now the UI displays the characters remaining moves and energy all of the time.
If a character runs out of energy -> all the tab is grayed out.
When the turn ends, all of it the tabs are grayed out.
At the start of new turn: alpha is set fully to 1 again.
End turn grays out all of the player unit profiles.